maParticles_Experiment_01_RadialEmitter.canvas
Test and experimentation canvas for initial release of the maParticles system. Simulation needs to get initialized on first frame (1.0), so if you are getting error messages because the canvas was stored at a later frame, please jump to first frame to initialize the simulation first.
The simulation node AND the radial emitter node both provide properties to play with.